+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-02T07:06:06.701Z"
+    created: Optional[str] = "2026-07-02T10:35:06.898Z"
 
-    updated: Optional[str] = "2026-07-02T07:06:06.701Z"
+    updated: Optional[str] = "2026-07-02T10:35:06.898Z"
 
     updated_by_id: OptionalNullable[str] = UNSET