+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWGTD7RX8DF4E9CJ7Y5K13TR"
+        "01KWH6BZ6SGPBW2J21RJQF57WE"
     )
     r"""The unique identifier of the data source"""