+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWGTD7RZHHN6H7H2N5DKDX1M"
+        "01KWH6BZ6VGEAF7R0GXS084F9R"
     )
     r"""The unique identifier of the data source"""