+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWGTD7S4DNTMPFBEJCGJE9HD"
+        "01KWH6BZ6Z7EZWCSR7S9TZNN2R"
     )
     r"""The unique identifier of the data source"""