+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KWGTD7M316TZMNBP68C2ZTTP"
+    id: Optional[str] = "01KWH6BZ3BKTCM2WFXFYBNNHZB"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWGTD7KS7VFAZD7KQY194FR7"
+        "tool_01KWH6BZ32Z4M1MGMMXC0JSB26"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KWGTD7KRHYBZ681BN679351P"
+    id: Optional[str] = "01KWH6BZ318HEGT2GQQYK4R5J6"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWGTD7KP31VF5HG2Q9QCZ3HC"
+        "tool_01KWH6BZ309D095YCST381WPVZ"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWGTD7KKWBMDN85PE72M0DNQ"
+        "tool_01KWH6BZ2YS665WJ0K9823TXHV"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWGTD7KGBAT1EG24KY64BW3F"
+        "tool_01KWH6BZ2WADMJXNKY7R7YBTTX"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWGTD7KE966509SF9EW9YGV2"
+        "tool_01KWH6BZ2TEBZT8SS94ZPJ405Y"
     )
 
     display_name: Optional[str] = None