+++ packages/orq-rc/RELEASES.md
@@ -9189,3 +9189,13 @@ Based on:
 - [python v4.11.0-rc.69] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.69] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.69 - packages/orq-rc
+
+## 2026-07-02 10:37:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.70] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.70] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.70 - packages/orq-rc